fix: fold agent-creator template parsing into agentCreatorFlow.ts to respect the frozen host-agent-import baseline
The previous commit put buildCreatorConfig in a new sibling file
(agentCreatorTemplates.ts), which meant the extension host imported a
second, new @agent/* deep-import specifier not present in
config/ratchets/host-agent-import-baseline.json — a ratchet CLAUDE.md
documents as frozen ("never widen a baseline"), so
hostAgentDeepImportRatchet.vitest.ts would fail on this diff. A bot
review on the PR (chatgpt-codex-connector) caught this correctly.

Fold the parsing/config-building code directly into
agentCreatorFlow.ts instead, which already defines CreatorConfig and
is the specifier the extension already imports — no new host-to-agent
edge, and the domain logic still sits beside the type it builds.
